Updated license header and did some more refactors
