+++ crates/supasim/src/lib.rs
@@ -30,7 +30,6 @@ use hal::{
 };
 use parking_lot::{Mutex, RwLock, RwLockReadGuard, RwLockWriteGuard};
 use smallvec::{SmallVec, smallvec};
-use std::collections::HashSet;
 use std::ops::Bound;
 use std::{
     marker::PhantomData,
@@ -39,7 +38,6 @@ use std::{
 };
 use thiserror::Error;
 use thunderdome::{Arena, Index};
-use types::SyncMode;
 
 pub use hal;
 pub use hal::{DeviceDescriptor, InstanceDescriptor};
@@ -49,7 +47,7 @@ pub use types::{
 };
 
 use crate::residency::BufferResidency;
-use crate::sync::DeviceSemaphore;
+use crate::sync::Semaphore;
 
 pub(crate) const DEVICE_SMALLVEC_SIZE: usize = 4;
 pub(crate) const STREAM_SMALLVEC_SIZE: usize = 16;
@@ -328,10 +326,8 @@ impl<B: hal::Backend> Instance<B> {
             myself: None,
         });
 
-        let handle = create_sync_thread(s.clone()).unwrap();
         {
             let mut inner_mut = s.inner_mut().unwrap();
-            inner_mut.sync_thread = Some(handle);
             inner_mut.myself = Some(s.downgrade());
         }
         s
@@ -430,127 +426,9 @@ impl<B: hal::Backend> Instance<B> {
     }
     pub fn submit_commands(
         &self,
-        recorders: &mut [crate::CommandRecorder<B>],
+        recorders: &[crate::CommandRecorder<B>],
     ) -> SupaSimResult<B, WaitHandle<B>> {
-        self.check_destroyed()?;
-        let submission_idx;
-        {
-            let s = self.inner()?;
-
-            let mut recorder_locks = Vec::new();
-            for r in recorders.iter_mut() {
-                r.check_destroyed()?;
-                recorder_locks.push(r.inner_mut()?);
-            }
-            let mut recorder_inners = Vec::new();
-            for r in &mut recorder_locks {
-                recorder_inners.push(&mut **r);
-            }
-
-            let mut recorder = if let Some(mut r) = s.hal_command_recorders.lock().pop() {
-                unsafe {
-                    r.clear(s.stream.lock().as_mut().unwrap()).map_supasim()?;
-                }
-                r
-            } else {
-                unsafe { s.stream.lock().as_mut().unwrap().create_recorder() }.map_supasim()?
-            };
-            let mut used_buffers = HashSet::new();
-            let mut used_buffer_ranges = Vec::new();
-            let streams = sync::assemble_streams(
-                &mut recorder_inners,
-                &s,
-                s.instance_properties.sync_mode == SyncMode::VulkanStyle,
-            )?;
-            for (buf_id, ranges) in &streams.used_ranges {
-                let b = s
-                    .buffers
-                    .lock()
-                    .get(*buf_id)
-                    .ok_or(SupaSimError::AlreadyDestroyed("Buffer".to_owned()))?
-                    .as_ref()
-                    .unwrap()
-                    .upgrade()?;
-                let _b = b.clone();
-                let b_mut = _b.inner()?;
-                for range in ranges {
-                    let id = b_mut.slice_tracker.acquire(
-                        &s,
-                        *range,
-                        if b_mut.slice_tracker.mutex.lock().gpu_available {
-                            BufferUser::Gpu(u64::MAX)
-                        } else {
-                            BufferUser::Cross(u64::MAX)
-                        },
-                        false,
-                    )?;
-                    used_buffer_ranges.push((id, b.clone()))
-                }
-                used_buffers.insert(buf_id);
-            }
-            drop(s);
-            let bind_groups = sync::record_command_streams(
-                &streams,
-                self.clone(),
-                &mut recorder,
-                &streams.resources.temp_copy_buffer,
-            )?;
-            let s = self.inner()?;
-            let used_buffers: Vec<_> = used_buffers
-                .iter()
-                .map(|a| {
-                    s.buffers
-                        .lock()
-                        .get(**a)
-                        .unwrap()
-                        .as_ref()
-                        .unwrap()
-                        .upgrade()
-                        .unwrap()
-                })
-                .collect();
-            let kernels = streams.resources.kernels.clone();
-            submission_idx = s.sync_thread().submit_gpu(GpuSubmissionInfo {
-                command_recorders: vec![recorder],
-                bind_groups,
-                used_buffer_ranges: used_buffer_ranges.clone(),
-                used_buffers,
-                used_resources: streams.resources,
-            })?;
-
-            for kernel in &kernels {
-                kernel.inner_mut()?.last_used = submission_idx;
-            }
-            for (buf_id, _) in streams.used_ranges {
-                let b = s
-                    .buffers
-                    .lock()
-                    .get(buf_id)
-                    .ok_or(SupaSimError::AlreadyDestroyed("Buffer".to_owned()))?
-                    .as_ref()
-                    .unwrap()
-                    .upgrade()?;
-                let mut b_mut = b.inner_mut()?;
-                b_mut.last_used = submission_idx;
-            }
-            for (id, b) in used_buffer_ranges {
-                b.inner()?
-                    .slice_tracker
-                    .update_user_submission(id, submission_idx, &s);
-            }
-        }
-
-        for recorder in recorders {
-            recorder._destroy()?;
-        }
-        Ok(WaitHandle::from_inner(WaitHandleInner {
-            _phantom: Default::default(),
-            _is_destroyed: false,
-            instance: self.clone(),
-            index: submission_idx,
-            id: Index::DANGLING,
-            is_alive: true,
-        }))
+        sync::submit_command_recorders(self, recorders)
     }
     pub fn wait_for_idle(&self, _timeout: f32) -> SupaSimResult<B, ()> {
         self.check_destroyed()?;
@@ -759,7 +637,7 @@ api_type!(Kernel, {
     instance: Instance<B>,
     per_device: SmallVec<[Option<B::Kernel>; DEVICE_SMALLVEC_SIZE]>,
     reflection_info: KernelReflectionInfo,
-    last_used_per_device: SmallVec<[Vec<DeviceSemaphore<B>>; DEVICE_SMALLVEC_SIZE]>,
+    last_used_per_device: SmallVec<[Vec<Semaphore<B>>; DEVICE_SMALLVEC_SIZE]>,
     id: Index,
 },);
 impl<B: hal::Backend> std::fmt::Debug for Kernel<B> {
@@ -1309,7 +1187,7 @@ impl<B: hal::Backend> Drop for MappedBuffer<'_, B> {
 
 api_type!(WaitHandle, {
     instance: Instance<B>,
-    semaphore: DeviceSemaphore<B>,
+    semaphore: Semaphore<B>,
     id: Index,
     is_alive: bool,
 },);
